+++ lib/mix/test/mix/tasks/compile.elixir_test.exs
@@ -96,22 +96,6 @@ defmodule Mix.Tasks.Compile.ElixirTest do
     end
   end
 
-  test "recompiles after path dependency changed" do
-    in_fixture("umbrella_dep/deps/umbrella/apps", fn ->
-      Mix.Project.in_project(:bar, "bar", fn _ ->
-        Mix.Tasks.Deps.Compile.run []
-
-        assert Mix.Tasks.Compile.Elixir.run([]) == :ok
-        assert Mix.Tasks.Compile.Elixir.run([]) == :noop
-        purge [Bar]
-
-        future = { { 2020, 4, 17 }, { 14, 0, 0 } }
-        File.touch!("../foo/ebin/.compile.elixir", future)
-        assert Mix.Tasks.Compile.Elixir.run([]) == :ok
-      end)
-    end)
-  end
-
   test "recompiles with --force" do
     in_fixture "no_mixfile", fn ->
       assert Mix.Tasks.Compile.Elixir.run([]) == :ok